fix folder scructure, add on + off methods to eventEmitter
